+++ packages/coding-agent/test/extensions-runner.test.ts
@@ -88,6 +88,7 @@ describe("ExtensionRunner", () => {
 		setActiveTools: () => {},
 		setDeferredOverrides: () => {},
 		setToolNamespaces: () => {},
+		hydrateMessageDeliveredTools: () => {},
 		refreshTools: () => {},
 		getCommands: () => [],
 		setModel: async () => false,